requirements.txt
openai
langchain
streamlit
pypdf
python-dotenv
pinecone-client==2.2.4
sentence_transformers
altair==4
tiktoken==0.4.0
chromadb==0.3.29